Move cmd/kyma to internal/cmd
