+++ internal/cmd/completion/completion.go
