+++ internal/cmd/install/cfg_darwin.go
