+++ internal/cmd/install/cfg_others.go
