+++ internal/cmd/install/cfg_windows.go
