+++ internal/cmd/install/cmd_test.go
