+++ internal/cmd/install/hosts_windows.go
