+++ internal/cmd/test/delete/cmd_test.go
