+++ internal/cmd/test/delete/opts.go
