+++ internal/cmd/test/status/cmd_test.go
