+++ internal/cmd/uninstall/cmd.go
