+++ internal/cmd/uninstall/cmd_test.go
