+++ internal/cmd/uninstall/opts.go
